backend/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# AI/ML dependencies
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.2
sentence-transformers>=2.2.2
transformers>=4.35.0
numpy>=1.24.0

# Vector database
qdrant-client>=1.7.0

# BM25 sparse retrieval
rank-bm25>=0.2.2

# Chinese text processing
jieba>=0.42.1

# Azure services
azure-cognitiveservices-speech>=1.34.0
pydub>=0.25.1

# Web search
google-api-python-client>=2.108.0
requests>=2.31.0

# Document processing
python-docx>=0.8.11
pypdf>=3.0.0
beautifulsoup4>=4.12.0
tiktoken>=0.5.0

# Environment and configuration
python-dotenv>=1.0.0

# Security and sanitization
bleach>=6.1.0

# Image processing (for medical image analysis)
Pillow>=10.1.0

# PyTorch for medical models (optional, for local models)
torch>=2.1.0
torchvision>=0.16.0
